remove chaos policy since it is now deprecated.
