+++ src/docs/guide/modules/chaos/example_chaos_controls_openapi.kt
@@ -1,8 +1,8 @@
 package guide.modules.chaos
 
 import org.http4k.chaos.ChaosBehaviours.ReturnStatus
-import org.http4k.chaos.ChaosPolicies.PercentageBased
-import org.http4k.chaos.inject
+import org.http4k.chaos.ChaosTriggers.PercentageBased
+import org.http4k.chaos.appliedWhen
 import org.http4k.chaos.withChaosControls
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.INTERNAL_SERVER_ERROR
@@ -16,7 +16,7 @@ import org.http4k.server.asServer
 fun main(args: Array<String>) {
     Cors(UnsafeGlobalPermissive)
             .then { Response(OK).body("A normal response") }
-            .withChaosControls(PercentageBased(50).inject(ReturnStatus(INTERNAL_SERVER_ERROR)))
+            .withChaosControls(ReturnStatus(INTERNAL_SERVER_ERROR).appliedWhen(PercentageBased(50)))
             .asServer(SunHttp(9000))
             .start()
             .also { println("Visit the app at http://localhost:9000 or see the OpenApi at https://www.http4k.org/swagger2/?url=http://localhost:9000/chaos")}
